pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nos"
version = "0.1.0"
description = "Workflow engine (library) and optional Flask + Socket.IO platform"
requires-python = ">=3.10"
# Minimal install: engine + schemas only. Platform: pip install 'nos[web]'. Everything: pip install 'nos[all]'.
dependencies = [
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
# Flask server, API, DB, Socket.IO, CLI — required to run `nos` and create_app().
web = [
    "psutil>=5.9.0",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "gevent>=23.9.0",
    "gevent-websocket>=0.10.1",
    "werkzeug>=2.3.0",
    "waitress>=2.1.0",
    "markdown>=3.4.0",
    "pygments>=2.15.0",
    "requests>=2.28.0",
]
# Optional node / integration stacks (not required for bare Workflow/Node usage).
integrations = [
    "twilio>=8.0.0",
    "ollama>=0.3.0",
    "chromadb>=0.4.0",
    "openpyxl>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.40.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
# Convenience: web + integrations + ai (duplicate lists; keep in sync when adding extras).
all = [
    "psutil>=5.9.0",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "gevent>=23.9.0",
    "gevent-websocket>=0.10.1",
    "werkzeug>=2.3.0",
    "waitress>=2.1.0",
    "markdown>=3.4.0",
    "pygments>=2.15.0",
    "requests>=2.28.0",
    "twilio>=8.0.0",
    "ollama>=0.3.0",
    "chromadb>=0.4.0",
    "openpyxl>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
]

[project.scripts]
nos = "nos.platform.cli:main"
nos-plugin = "nos.platform.plugins.cli:main"

# Engine plugins discoverable at startup (Node / Workflow / optional Plugin subclasses).
[project.entry-points."nos.plugins"]
test_chain_math = "nos.plugins.workflows.test.chain_math_workflow:ChainMathWorkflow"
simple_sum_node = "nos.plugins.nodes.test.simple_sum_node:SimpleSumNode"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
